Add memory handle to Creeps
Also introduced observer/observable scheme and WeakRef for scripting
environments.  Now the RoomObjects can be safely destroyed even if
the VMs have reference to them.

Wren has a problem that the references will be kept until GC runs, whose
timing is totally unpredictable.
Squirrel is preferable because it finalizes the references as soon as
there are no reference to them, which is achieved by reference counting
at the small cost of performance.  That said, keeping long lists of
Observers might beat the performance gain of not reference counting.